+++ src/main/java/com/junkai/picture_enhancement_platform/controller/PictureController.java
@@ -18,7 +18,6 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.io.IOException;
 
 @RestController
 @RequestMapping("/image")
