+++ src/main/java/com/junkai/picture_enhancement_platform/service/interfaces/ModelService.java
@@ -1,7 +1,6 @@
 package com.junkai.picture_enhancement_platform.service.interfaces;
 
 import com.junkai.picture_enhancement_platform.POJO.ModelParameterEntity;
-import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
