+++ src/main/java/com/junkai/picture_enhancement_platform/service/interfaces/UserService.java
@@ -2,7 +2,6 @@ package com.junkai.picture_enhancement_platform.service.interfaces;
 
 import com.baomidou.mybatisplus.extension.service.IService;
 import com.junkai.picture_enhancement_platform.entity.User;
-import org.springframework.stereotype.Service;
 
 
 public interface UserService extends IService<User> {
